feat(inspection): 实现数据库配置与实体映射优化
- 添加 Hibernate 配置初始化逻辑，支持从配置文件和 ConfigManager 加载设置
- 实现自定义物理命名策略，支持驼峰与下划线转换- 更新实体类命名与映射关系，将 PO 后缀统一为 DO-优化数据库服务初始化流程，移除冗余配置代码
- 调整缓存配置引用类型，确保使用最新的实体类- 添加数据库查询测试类，验证连接与基本操作-修复模板信息查询方法返回类型不一致问题
- 更新应用配置文件，补充 PostgreSQL 连接参数
- 移除旧版 Hibernate 配置加载逻辑，统一使用 AppConfig 管理
- 调整实体间关联关系，明确外键映射与级联操作
- 修正字段类型定义，提升数据映射准确性
- 添加事务控制与异常处理机制，增强数据操作稳定性

+++ qtech/qtech-common/src/main/java/com/im/qtech/common/dto/standard/EqLstTplInfo.java
@@ -18,11 +18,15 @@ public class EqLstTplInfo implements Serializable {
     private static final long serialVersionUID = 529L;
 
     private String module;
-    private Integer listParams;
-    private Integer itemParams;
+    private Long listParams;
+    private Long itemParams;
     private Integer status;
 
-    // 重写equals和hashCode方法，用于判断对象的对应属性是否相等
+    /**
+     * 重写equals和hashCode方法，用于判断对象的对应属性是否相等
+     * 重写 equals/hashCode，使用业务字段，而非数据库主键
+     * 注意：如果你想用在 Set / Map 做去重，业务字段足够，避免使用懒加载对象 template
+     */
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
